+++ wrapper/src/wrapper.cpp
@@ -318,6 +318,13 @@ void LLVMSetDSOLocal(LLVMValueRef Global, bool value)
 	llvm::unwrap<llvm::GlobalValue>(Global)->setDSOLocal(value);
 }
 
+void LLVMSetNoSanitizeAddress(LLVMValueRef Global)
+{
+	auto global = llvm::unwrap<llvm::GlobalValue>(Global);
+	auto data = global->getSanitizerMetadata();
+	data.NoAddress = 1;
+	global->setSanitizerMetadata(data);
+}
 void LLVMBuilderSetFastMathFlags(LLVMBuilderRef Builder, FastMathOption option)
 {
 	llvm::FastMathFlags math_flags {};